# backend/assets/types/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pontiac API",
        "description": "The JSON API behind the React UI. Routes, parameters and \u2014 for endpoints that return a response DTO \u2014 the response shapes.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/platform/accounts/{id}": {
            "get": {
                "operationId": "get_api_platform_account_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountDetailOutput"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "summary": "Datos, and L\u00edmites y funciones. A field left out is left as it is.",
                "operationId": "patch_api_platform_account_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountDetailOutput"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/platform/accounts/{id}/suspend": {
            "post": {
                "summary": "Its people cannot sign in and its public pages are not served, until it is reactivated. Nothing is deleted.",
                "operationId": "post_api_platform_account_suspend",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountDetailOutput"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/platform/accounts/{id}/reactivate": {
            "post": {
                "operationId": "post_api_platform_account_reactivate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountDetailOutput"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/platform/accounts/{id}/users": {
            "get": {
                "summary": "The owner and the assistants. ?q= searches name and email. Clients are the consultant's, not listed here.",
                "operationId": "get_api_platform_account_users",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "items",
                                        "total",
                                        "page",
                                        "perPage"
                                    ],
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TeamMemberOutput"
                                            }
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "page": {
                                            "type": "integer"
                                        },
                                        "perPage": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/platform/accounts/{id}/users/{userId}/resend-invitation": {
            "post": {
                "operationId": "post_api_platform_account_user_resend",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamMemberOutput"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/platform/accounts/{id}/users/{userId}": {
            "delete": {
                "summary": "Disables an assistant. The owner is not disabled: suspend the consultant instead.",
                "operationId": "delete_api_platform_account_user_disable",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamMemberOutput"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/platform/accounts/{id}/users/{userId}/enable": {
            "post": {
                "summary": "The super admin may enable an assistant beyond the consultant's limit: the limit binds the consultant.",
                "operationId": "post_api_platform_account_user_enable",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamMemberOutput"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "TeamMemberOutput": {
                "required": [
                    "id",
                    "email",
                    "fullName",
                    "role",
                    "active",
                    "loginStatus"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "role": {
                        "title": "\"owner\" or \"assistant\".",
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "loginStatus": {
                        "title": "\"active\" (has a password), \"invited\" (link sent, not used yet) or \"none\".",
                        "type": "string"
                    },
                    "lastSignInAt": {
                        "title": "ISO 8601; null until they first sign in.",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "AccountOwnerOutput": {
                "required": [
                    "id",
                    "fullName",
                    "email",
                    "loginStatus"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "loginStatus": {
                        "title": "\"active\" (has a password), \"invited\" or \"none\".",
                        "type": "string"
                    },
                    "lastSignInAt": {
                        "title": "ISO 8601; null until they first sign in.",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "AccountDetailOutput": {
                "required": [
                    "id",
                    "name",
                    "slug",
                    "country",
                    "currency",
                    "locale",
                    "timezone",
                    "active",
                    "createdAt",
                    "maxPublishedPages",
                    "maxAssistants",
                    "storageMb",
                    "maxFileMb",
                    "features",
                    "assistants",
                    "clients"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "locale": {
                        "type": "string"
                    },
                    "timezone": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "title": "ISO 8601.",
                        "type": "string"
                    },
                    "maxPublishedPages": {
                        "type": "integer"
                    },
                    "maxAssistants": {
                        "type": "integer"
                    },
                    "storageMb": {
                        "type": "integer"
                    },
                    "maxFileMb": {
                        "type": "integer"
                    },
                    "features": {
                        "title": "the features that are on: \"booking\", \"payments\", \"portal\", \"flows\"",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "owner": {
                        "nullable": true,
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/AccountOwnerOutput"
                            }
                        ]
                    },
                    "assistants": {
                        "type": "integer"
                    },
                    "clients": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}